Update Fitting_testing_MvD_LBD.py
adding 1st function of data_spectra_adder_merger_thingy. The base is a singe file with all spectra in it as seperate columns. The first is the energy, the others are each spectra to fit.
NOTE: to the energy 10000 is always added to differentiate each spectra (so the 0th(nr1) is the base BE, the 1st has the "prevalue" 10xxx, the seccond 20xxx etc). This was choosen to make it easier for the distinction of BE for the centers in the parameter_file for each spectra (updated next)
Here is it important, that the header (E, S00, S01..) is skipped! --> skip_rows needs to entered as 1. this will be asked in a different function next aswell