Runtime improvement + data_output
With this update the fitting runtime was greatly improved. From over 115 days down to 1!!! (with my 169+18 peaks)
Also now the result out will be written into a file

Main reason for the time saving:
Param_updater.py: new kind of sorting and searching though all of it
--> 115d -> 3.5d

fitting fkt.py: better initilizing of the np.array, also not always using f"..." but using an new var for this.
--> 3.5d->~1

Plotting_functions.py: delete obsolete function, which was written already in other file (not used for fitting but afterwards)