corelevl splitting differentiation, Inputs-file & 2D plot Inputs-file Update
with this update an update was done regarding :

1.) Analysis_ana-py:
the calculations of oxid states. Now one can assign the peaks to the oxid state an swell as to the splitting it belongs to (in the Inputs_file). for e.g.2p:  3/2 or 1/2 etc.  This now also has an impact on the calculations since now only the 3/2 is used and not, as before the full spectra.

2.) Analysis_plotting:
The 2d plot was updated
Also a small iteration was written, to include the old shirely BG (used for Ni2p, only used in my case)